+++ frontend/src/components/UI/SummaryCard/CurrencySummaryCard/CurrencySummaryCard.jsx
