+++ frontend/src/components/UI/SummaryCard/CurrencySummaryCard/CurrencySummaryCard.test.js
